fix tests to work wth css modules for AddLead
