+++ client/package.json
@@ -69,6 +69,7 @@
     "chai": "^4.1.2",
     "enzyme": "^3.4.1",
     "enzyme-adapter-react-16": "^1.2.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest-localstorage-mock": "^2.2.0",
     "jsdom": "11.12.0",
     "jsdom-global": "3.0.2"
@@ -97,7 +98,9 @@
       "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$"
     ],
     "moduleNameMapper": {
-      "^react-native$": "react-native-web"
+      "^react-native$": "react-native-web",
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less)$": "identity-obj-proxy"
     },
     "moduleFileExtensions": [
       "web.js",
